pyflask/day10/gradepm/requirements.txt
flask
flask-sqlalchemy
flask-script
flask-migrate
flask-restful
flask-blueprint
pymysql